bytes_terminate(): speed up by using `String#index`
Based on the current implementation of `bytes_terminate` in our Python
runtime lib:
https://github.com/kaitai-io/kaitai_struct_python_runtime/blob/07aea9c6cdb1cc5be8677004680382602d7323f3/kaitaistruct.py#L472-L477

Using the built-in
[`String#index`](https://docs.ruby-lang.org/en/3.3/String.html#method-i-index)
method (implemented in C) instead of our pure Ruby implementation
simplifies the code and significantly improves performance, see
https://gist.github.com/generalmimon/331361de0ee72d4ab1f952ea92c8bec7#file-results-index-vs-old-sh-session

As you can see, the new implementation is generally two orders of
magnitude faster than the old one.

This is very similar to the change that was made in Python. In
https://github.com/kaitai-io/kaitai_struct_python_runtime/pull/42, the
manual implementation of `bytes_terminate` was replaced by the native
`bytes.partition` method in Python. Later it was found that the
`partition` method is actually suboptimal for our use in
`bytes_terminate` and that it's faster to use `bytes.index` + slicing
(see
https://github.com/kaitai-io/kaitai_struct_python_runtime/commit/92a2d71519ba1c99390f2755ab0ec0c635176f40),
but even the version with `partition` was already much faster than our
previous pure Python version, see
https://github.com/kaitai-io/kaitai_struct_python_runtime/pull/42#issuecomment-716215964