substream(): ensure identical behavior to read_bytes()
Fixes the {Eof,Eos}ExceptionSized tests in Ruby (see
https://github.com/kaitai-io/ci_artifacts/blob/d7568fe794b09292b98873dcdc1a5534928345b6/test_out/ruby/ci.json#L314-L324)

Overall, this commit tries to ensure that the zero-copy substreams
(https://github.com/kaitai-io/kaitai_struct/issues/44) are a drop-in
replacement for the old way using `read_bytes` + `Stream#new` without
any observable differences. It's not perfect, but it should be close
enough.

To verify this, we automatically generate and evaluate 2000 random
scenarios using https://github.com/rantly-rb/rantly involving 10 stream
operations and compare whether the old and new substream implementations
behave the same. If not, the test will fail. It is a normal RSpec test,
so it can be run along with other unit tests using `rspec spec/` on the
command line.

Every single change made to the stream implementation follows from a
particular edge case identified by this extensive testing.