GH Actions: split RuboCop into warnings and "conventions"
Currently, RuboCop is reporting many findings, but most of them have a
severity of `[C] convention`, which are not nearly as important as
warnings (`[W] warning`). Therefore, we split warnings and conventions
using the following option (this is an explanation from `rubocop
--help`):

> `--display-only-fail-level-offenses`: Only output offense messages at
> the specified `--fail-level` or above.

In addition, this means that any warnings will mark the CI workflow as
"failed", whereas conventions will not.