read_{s,u}{2,4,8}{be,le}(): use proper unpacking directives
This simplifies the code a bit.

See https://docs.ruby-lang.org/en/3.4/packed_data_rdoc.html#label-For+Integers

Strictly speaking, this raises the minimum required Ruby version to
exactly 1.9.3. This is because the `>` and `<` directives were added in
Ruby 1.9.3, see the note "(available since Ruby 1.9.3)" at
https://ruby-doc.org/core-1.9.3/String.html#method-i-unpack. As
expected, there is no mention of them in the Ruby 1.9.2 documentation:
https://ruby-doc.org/core-1.9.2/String.html#method-i-unpack

I don't think anyone cares because Ruby 1.9.3 was released in October
2011 (13 years ago), but let's still specify it as the minimum required
Ruby version in .gemspec.